many style adjustments
