+++ src/front/js/component/cityCard.js
@@ -5,19 +5,17 @@ import { Link } from "react-router-dom";
 
 const CityCard = ({ name, text1, text2, text3, text4, text5 }) => {
   return (
-    <div className="row">
-      <div className="col-sm-6">
-        <div className="card text-light bg-dark">
-          <div className="card-body">
-            <h5 className="card-title">
-              <strong>{name ? name : "Filter titles"}</strong>
-            </h5>
-            <p className="card-text">{text1 ? text1 : "Filter titles"}</p>
-            <p className="card-text">{text2 ? text2 : "Filter titles"}</p>
-            <p className="card-text">{text3 ? text3 : "Filter titles"}</p>
-            <p className="card-text">{text4 ? text4 : "Filter titles"}</p>
-            <p className="card-text">{text5 ? text5 : "Filter titles"}</p>
-          </div>
+    <div className="col flex-fill">
+      <div className="card text-dark bg-light mx-3">
+        <div className="card-title">
+          <strong>{name ? name : "Filter titles"}</strong>
+        </div>
+        <div className="card-body">
+          <p className="card-text">{text1 ? text1 : "Filter titles"}</p>
+          <p className="card-text">{text2 ? text2 : "Filter titles"}</p>
+          <p className="card-text">{text3 ? text3 : "Filter titles"}</p>
+          <p className="card-text">{text4 ? text4 : "Filter titles"}</p>
+          <p className="card-text">{text5 ? text5 : "Filter titles"}</p>
         </div>
       </div>
     </div>
